requirements.txt
kafka-python
scikit-fuzzy
pandas
scipy
packaging
numpy
networkx
